optimize routing module by lazy loading
