+++ CSETWebNg/src/app/routing/reports-routing/edm-routing.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { EdmComponent } from '../../reports/edm/edm.component';
+
+const routes: Routes = [
+    {
+        path: '',
+        component: EdmComponent,
+    }
+];
+
+@NgModule({
+    imports: [RouterModule.forChild(routes)],
+    exports: [RouterModule]
+})
+export class EdmRoutingModule { }
